# contracts/openapi.yaml
...
openapi: 3.1.0
info:
  title: Curiouskids Club API
  version: 1.0.0
  description: Lending library API. Errors are RFC 9457 problem+json.
paths:
...
  /api/v1/auth/login:
    post:
      tags:
        - auth
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: Logged in; the session cookie is set
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          description: 'UNAUTHENTICATED: email or password is incorrect'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '429':
          description: 'RATE_LIMITED: too many failed attempts; see Retry-After'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
...
  /api/v1/auth/me:
    get:
      tags:
        - auth
      operationId: me
      responses:
        '200':
          description: The logged-in account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          description: 'UNAUTHENTICATED: not logged in'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
...
components:
  schemas:
    Problem:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
          format: int32
        code:
          type: string
          enum:
            - VALIDATION_FAILED
            - UNAUTHENTICATED
            - FORBIDDEN
            - NOT_FOUND
            - METHOD_NOT_ALLOWED
            - NOT_ACCEPTABLE
            - CONFLICT
            - UNSUPPORTED_MEDIA_TYPE
            - EMAIL_NOT_VERIFIED
            - SLOT_REQUIRED
            - SLOT_FULL
            - SLOT_NOT_BOOKABLE
            - LIMIT_REACHED
            - ALREADY_HOLDING_TITLE
            - COPY_NOT_AVAILABLE
            - RENEWAL_NOT_ALLOWED
            - TOKEN_INVALID
            - IDEMPOTENCY_KEY_REUSED
            - RATE_LIMITED
            - INTERNAL_ERROR
        detail:
          type: string
        traceId:
          type: string
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ProblemFieldError"
      required:
        - code
        - status
        - title
        - traceId
        - type
    ProblemFieldError:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
      required:
        - field
        - message
...
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          maxLength: 254
          minLength: 1
        password:
          type: string
          maxLength: 128
          minLength: 1
      required:
        - email
        - password
    MeResponse:
      type: object
      properties:
        accountId:
          type: string
          format: uuid
        role:
          type: string
          enum:
            - MEMBER
            - VOLUNTEER
            - ADMIN
        familyId:
          type: string
          format: uuid
          description: Absent for staff
        emailVerified:
          type: boolean
      required:
        - accountId
        - emailVerified
        - role
...